% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csDAM.R
\name{select_genes_by_k}
\alias{select_genes_by_k}
\title{Select features by k threshold}
\usage{
select_genes_by_k(data, cos_mat, k_thr = NULL)
}
\arguments{
\item{data}{A gene expression matrix (rows = genes, columns = samples).}

\item{cos_mat}{A cosine similarity matrix (rows = genes, columns = samples).}

\item{k_thr}{Threshold of k for selecting genes. If NULL (default), it will
be set to the number of samples (i.e., ncol(data)).}
}
\value{
A list containing:
\item{selected_genes}{Names of selected genes.}
\item{best_k}{Vector of best-k indices for selected genes.}
\item{pi}{A list of permutations (descending order of expression across samples).}
}
\description{
Select genes by k threshold
}
\details{
This function selects genes based on the maximum cosine similarity index
and orders their expression across samples.

Optionally it rules out features that are unlikely to be the markers, e.g.,
those features that resemble 1,1,...,1,1 than any other reference vectors.
}
\examples{
\dontrun{
res <- select_genes_by_k(data, cos_mat)
}

}
